Refector to simplify API
