+++ test/test_rules.py
@@ -1,5 +1,5 @@
 from opimodel import rules
-from cssgen import render
+from renderers import render
 
 
 def test_empty_RulesNode(widget, get_renderer):
